Fix issue caused by duplicated main adapters causing crash:
Issue: First adapter is created which creates all the fragments. Dining fragment spawns new thread for network request. Then, startHomeFragment() is called in MainActivity, which creates a new adapter and destroys the old one. Then, the dining network request finishes and tries to access a null binding causing crash.

Added Crashyltics error logging for when this happens but prevents crash on the User side.